backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Data validation
pydantic==2.9.2
pydantic[email]==2.9.2

# CORS (included in FastAPI but explicit)
# Already included in fastapi
